LayoutTests/fast/viewport/viewport-35.html

<head>
    <title>Viewport META with "width=480, initial-scale=0.1".</title>
    <meta name="viewport" content="width=480, initial-scale=0.1">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
