Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-35-expected.txt

ALERT: viewport size 1280x1408 scale 0.25 with limits [0.25, 5] and userScalable true
